+++ frontend/src/login.js
@@ -1,0 +1,28 @@
+import React from 'react';
+
+const login = () => {
+	return (
+		<>
+			<div className="Title">
+				<h1>Embry-Riddle Aeronautical University</h1>
+				<h2>Faculty Information Repository</h2>
+			</div>
+	
+			<div>
+				<center><table className = "LoginStatement">
+					<tr>
+						<th><h3>Server Login</h3></th>
+					</tr>
+					<tr>
+						<th><h1>TextFieldGoesHere</h1></th>
+					</tr>
+					<tr>
+						<th><h4>Contact Department Head with issues</h4></th>
+					</tr>
+				</table></center>
+			</div>
+		</>
+	);
+};
+
+export default login;
