--- src/pages/NoPage.js
@@ -1,5 +1,0 @@
-const NoPage = () => {
-  return <h1>404</h1>;
-};
-
-export default NoPage;
